Completed Navbar as in the wireframe and created seperate components for navbar and sidebar
